requirements.txt
fastapi
aiogram>=3
gunicorn
httpx
python-dotenv
uvicorn
